ct_dicom/requirements.txt
absl-py
file-io
numpy
pydicom>=2.3.1
pypng
tensorflow-cpu
